[4] Split into separate crates huginn-net-tcp  (#176)
* init tcp module

* fmt fixed

* [5] Split into separate crates huginn-net-core  (#177)

* init core

* output split

* update ci by workspaces

* documented code update
