+++ .github/workflows/audit.yml
@@ -5,6 +5,7 @@ on:
     branches: [ "master" ]
     paths:
       - '**/Cargo.toml'
+      - '**/Cargo.lock'
   schedule:
     - cron: '0 2 * * *'
 
@@ -13,6 +14,7 @@ permissions:
 
 jobs:
   cargo-deny:
+    name: Cargo Deny (Workspace)
     permissions:
       checks: write
       contents: read
@@ -23,6 +25,7 @@ jobs:
     - uses: EmbarkStudios/cargo-deny-action@v2
 
   cargo-audit:
+    name: Cargo Audit (Workspace)
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v5
