+++ examples/capture.rs
@@ -1,5 +1,5 @@
 use clap::{Parser, Subcommand};
-use huginn_net::fingerprint_result::FingerprintResult;
+use huginn_net::output::FingerprintResult;
 use huginn_net::Database;
 use huginn_net::HuginnNet;
 use std::sync::atomic::{AtomicBool, Ordering};
