+++ huginn-net/src/error.rs
